extracted new component for managing fastas
